Rewrite Gps location acquisition
